+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.9.1 (unreleased)
 
-* Backport rack2 compatibility from #1260
+* Backport rack2 compatibility (#1260)
+
 * Fixed issue with JS nil return paths being treated as true (#1274)
+
+* Fix using more than two `rescue` in sequence (#1269)
 
 
 ## 0.9.0 2015-12-20